+++ dev/openchpl_data-model.sql
@@ -65,8 +65,9 @@ CREATE TABLE openchpl.user(
 	credentials_expired bool NOT NULL,
 	account_enabled bool NOT NULL,
 	compliance_signature timestamp,
-        password_reset_required bool not null default false,
+    password_reset_required bool not null default false,
 	failed_login_count int not null default 0,
+    last_logged_in_date timestamp,
 	creation_date timestamp NOT NULL DEFAULT NOW(),
 	last_modified_date timestamp NOT NULL DEFAULT NOW(),
 	last_modified_user bigint NOT NULL,
@@ -2770,6 +2771,33 @@ CREATE TABLE openchpl.broken_surveillance_rules
     CONSTRAINT broken_surveillance_rules_id_pk PRIMARY KEY (id)
 );
 
+CREATE TABLE openchpl.url_type (
+	id bigserial NOT NULL,
+	name varchar(50) NOT NULL,
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT url_type_pk PRIMARY KEY (id)
+);
+
+CREATE TABLE openchpl.url_check_result (
+	id bigserial NOT NULL,
+	url_type_id bigint NOT NULL,
+	url text NOT NULL,
+	response_code int, --allow null in case something times out?
+	response_message text, -- if there is some additional text about the response (error message?)
+	checked_date timestamp, -- null to indicate we know about the URL in the system but haven't checked it yet
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT url_check_result_pk PRIMARY KEY (id),
+	CONSTRAINT url_type_fk FOREIGN KEY (url_type_id)
+		REFERENCES openchpl.url_type (id)
+		MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT
+);
+
 CREATE TABLE openchpl.test_functionality_criteria_map
 (
     id bigserial NOT NULL,
@@ -3113,6 +3141,80 @@ CREATE TABLE openchpl.quarterly_report_surveillance_map (
 	CONSTRAINT surveillance_process_type_fk FOREIGN KEY (surveillance_process_type_id)
 		REFERENCES openchpl.surveillance_process_type (id) 
 		MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT
+);
+
+CREATE TABLE openchpl.change_request_type (
+    id bigserial NOT NULL,
+    name text NOT NULL,
+    creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT change_request_type_pk PRIMARY KEY (id)
+);
+
+CREATE TABLE openchpl.change_request_status_type (
+    id bigserial NOT NULL,
+    name text NOT NULL,
+    creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT change_request_status_type_pk PRIMARY KEY (id)
+);
+
+CREATE TABLE openchpl.change_request (
+    id bigserial NOT NULL,
+    change_request_type_id bigint NOT NULL,
+    developer_id bigint NOT NULL,
+    creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT change_request_pk PRIMARY KEY (id),
+    CONSTRAINT change_request_type_fk FOREIGN KEY (change_request_type_id)
+	    REFERENCES openchpl.change_request_type (id)
+        MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT,
+    CONSTRAINT developer_fk FOREIGN KEY (developer_id)
+        REFERENCES openchpl.vendor (vendor_id)
+        MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT
+);
+
+CREATE TABLE openchpl.change_request_status (
+    id bigserial NOT NULL,
+    change_request_id bigint NOT NULL,
+    change_request_status_type_id bigint NOT NULL,
+    status_change_date timestamp NOT NULL,
+    comment text,
+    certification_body_id bigint,
+    creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT change_request_status_pk PRIMARY KEY (id),
+    CONSTRAINT change_request_fk FOREIGN KEY (change_request_id)
+	    REFERENCES openchpl.change_request (id)
+        MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT,
+    CONSTRAINT change_request_status_type_fk FOREIGN KEY (change_request_status_type_id)
+	    REFERENCES openchpl.change_request_status_type (id)
+        MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT,
+    CONSTRAINT certification_body_fk FOREIGN KEY (certification_body_id)
+	    REFERENCES openchpl.certification_body (certification_body_id)
+        MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT
+);
+
+CREATE TABLE openchpl.change_request_website (
+    id bigserial NOT NULL,
+    change_request_id bigint NOT NULL,
+    website text,
+    creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT change_request_website_pk PRIMARY KEY (id),
+    CONSTRAINT change_request_fk FOREIGN KEY (change_request_id)
+	    REFERENCES openchpl.change_request (id)
+        MATCH SIMPLE ON UPDATE NO ACTION ON DELETE RESTRICT
 );
 
 CREATE INDEX fki_certified_product_id_fk