+++ dev/openchpl_views.sql
@@ -11,6 +11,7 @@ DROP VIEW IF EXISTS openchpl.product_active_owner_history_map;
 DROP VIEW IF EXISTS openchpl.certified_product_summary;
 DROP VIEW IF EXISTS openchpl.ehr_certification_ids_and_products;
 DROP VIEW IF EXISTS openchpl.listings_from_banned_developers;
+DROP VIEW IF EXISTS openchpl.surveillance_basic;
 
 create or replace function openchpl.get_testing_lab_code(input_id bigint) returns
     table (
@@ -480,6 +481,38 @@ CREATE VIEW openchpl.certified_product_details AS
                   WHERE surv_1.deleted <> true AND nc_status.name::text = 'Closed'::text) n_1(id, certified_product_id, friendly_id, start_date, end_date, type_id, randomized_sites_used, creation_date, last_modified_date, last_modified_user, deleted, user_permission_id, id_1, surveillance_id, type_id_1, certification_criterion_id, requirement, result_id, creation_date_1, last_modified_date_1, last_modified_user_1, deleted_1, id_2, surveillance_requirement_id, certification_criterion_id_1, nonconformity_type, nonconformity_status_id, date_of_determination, corrective_action_plan_approval_date, corrective_action_start_date, corrective_action_must_complete_date, corrective_action_end_date, summary, findings, sites_passed, total_sites, developer_explanation, resolution, creation_date_2, last_modified_date_2, last_modified_user_2, deleted_2, id_3, name, creation_date_3, last_modified_date_3, last_modified_user_3, deleted_3)
           GROUP BY n_1.certified_product_id) nc_closed ON a.certified_product_id = nc_closed.certified_product_id;
 -- ALTER VIEW openchpl.certified_product_details OWNER TO openchpl;
+
+CREATE VIEW openchpl.surveillance_basic AS
+SELECT 
+	surv.*,
+	COALESCE(nc_open.count_open_nonconformities, 0::bigint) AS open_nonconformity_count,
+    COALESCE(nc_closed.count_closed_nonconformities, 0::bigint) AS closed_nonconformity_count
+FROM openchpl.surveillance surv
+LEFT JOIN
+  (SELECT surv.id AS surv_id,
+          count(*) AS count_open_nonconformities
+   FROM openchpl.surveillance surv
+   JOIN openchpl.surveillance_requirement surv_req ON surv.id = surv_req.surveillance_id
+   AND surv_req.deleted <> TRUE
+   JOIN openchpl.surveillance_nonconformity surv_nc ON surv_req.id = surv_nc.surveillance_requirement_id
+   AND surv_nc.deleted <> TRUE
+   JOIN openchpl.nonconformity_status nc_status ON surv_nc.nonconformity_status_id = nc_status.id
+   WHERE surv.deleted <> TRUE
+     AND nc_status.name::text = 'Open'::text
+   GROUP BY surv.id) nc_open ON surv.id = nc_open.surv_id
+LEFT JOIN
+  (SELECT surv.id AS surv_id,
+          count(*) AS count_closed_nonconformities
+   FROM openchpl.surveillance surv
+   JOIN openchpl.surveillance_requirement surv_req ON surv.id = surv_req.surveillance_id
+   AND surv_req.deleted <> TRUE
+   JOIN openchpl.surveillance_nonconformity surv_nc ON surv_req.id = surv_nc.surveillance_requirement_id
+   AND surv_nc.deleted <> TRUE
+   JOIN openchpl.nonconformity_status nc_status ON surv_nc.nonconformity_status_id = nc_status.id
+   WHERE surv.deleted <> TRUE
+     AND nc_status.name::text = 'Closed'::text
+   GROUP BY surv.id) nc_closed ON surv.id = nc_closed.surv_id
+WHERE surv.deleted = false;
 
 CREATE VIEW openchpl.certified_product_search AS
 SELECT cp.certified_product_id,
@@ -971,10 +1004,10 @@ WHERE deleted = false;
 
 CREATE VIEW openchpl.certified_product_summary AS
  SELECT cp.certified_product_id,
+	(select chpl_product_number from openchpl.get_chpl_product_number(cp.certified_product_id)),
     cp.certification_edition_id,
     cp.product_version_id,
     cp.certification_body_id,
-    cp.chpl_product_number,
     cp.report_file_location,
     cp.sed_report_file_location,
     cp.sed_intended_user_description,
@@ -1004,12 +1037,23 @@ CREATE VIEW openchpl.certified_product_summary AS
     p.name AS product_name,
     v.name AS vendor_name,
     v.vendor_code,
+	contact.full_name,
+	contact.email,
+	contact.phone_number,
+	pv.version,
     lastCertStatusEvent.certification_status,
+	certStatusEvent.certification_date,
     cb.acb_code,
     cb.name AS certification_body_name,
     cb.website AS certification_body_website
    FROM openchpl.certified_product cp
      JOIN openchpl.certification_edition ce ON cp.certification_edition_id = ce.certification_edition_id
+	 JOIN (
+		SELECT MIN(event_date) as "certification_date", certified_product_id 
+		FROM openchpl.certification_status_event 
+		WHERE certification_status_id = 1 
+		GROUP BY (certified_product_id)) certStatusEvent 
+		ON cp.certified_product_id = certStatusEvent.certified_product_id
 	 JOIN (
 		SELECT certStatus.certification_status as "certification_status", cse.certified_product_id as "certified_product_id"
 		FROM openchpl.certification_status_event cse
@@ -1026,6 +1070,12 @@ CREATE VIEW openchpl.certified_product_summary AS
      JOIN openchpl.product_version pv ON cp.product_version_id = pv.product_version_id
      JOIN openchpl.product p ON pv.product_id = p.product_id
      JOIN openchpl.vendor v ON p.vendor_id = v.vendor_id
+	 LEFT OUTER JOIN ( SELECT contact.contact_id,
+            contact.full_name,
+            contact.email,
+            contact.phone_number,
+            contact.title
+           FROM openchpl.contact) contact ON v.contact_id = contact.contact_id
      JOIN openchpl.certification_body cb ON cp.certification_body_id = cb.certification_body_id
 	 LEFT OUTER JOIN ( SELECT muu.meaningful_use_users,
             muu.certified_product_id,