+++ RELEASE_NOTES.md
@@ -1,5 +1,39 @@
 # Release Notes
 
+## Version 17.12.0
+_7 October 2019_
+
+### Features
+* Add tables and views to support reporting on questionable urls
+  * Add url_check_result table to store all the results of all urls that have been checked
+  * Add url_type table as a lookup table of places urls can be found in the system
+  * Update certified_product_summary view to have developer contact info and certification date
+
+### Data Changes
+* Add new RT 13, 14, and 15 Macra Measures
+
+---
+
+## Version 17.11.0
+_23 September 2019_
+
+### Features
+* Add tables to support persisting Website Change Requests
+
+---
+
+## Version 17.10.0
+_10 September 2019_
+
+### Features
+* Support storing the last logged in date with the user
+* Add surveillance basic view to include open/closed nonconformity count
+
+### Data Changes
+* Add new filter type for API Key Usage report
+
+---
+
 ## Version 17.9.0
 _26 August 2019_